Refactor: Divide ProjectsController into ProjectsAdminController, ProjectsUserController, and ProjectsController to improve separation of concerns and maintainability.
