+++ StudentProjectsCenterSystem.Core/Entities/DTO/Project/ProjectCreateDTO.cs
@@ -1,9 +1,14 @@
-﻿namespace StudentProjectsCenterSystem.Core.Entities.DTO.Project
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace StudentProjectsCenterSystem.Core.Entities.DTO.Project
 {
     public class ProjectCreateDTO
     {
+        [Required]
         public string Name { get; set; } = string.Empty;
+        [Required]
         public string SupervisorId { get; set; } = string.Empty;
+        [Required]
         public string CustomerId { get; set; } = string.Empty;
     }
 }
